+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Development]
+Nothing yet!
+
+## [v2.1.0]
 ### Added
 - The `autoExtendFail` event, fired when the TimeoutExtender fails to change the VisibilityTimeout of a message due to the message no longer existing or being expired. This event may encompass more errors in the future.
 
@@ -110,7 +113,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release
 
-[Development]: https://github.com/TechnologyAdvice/Squiss/compare/v2.0.2...HEAD
+[Development]: https://github.com/TechnologyAdvice/Squiss/compare/v2.1.0...HEAD
+[v2.1.0]: https://github.com/TechnologyAdvice/Squiss/compare/v2.0.2...v2.1.0
 [v2.0.2]: https://github.com/TechnologyAdvice/Squiss/compare/v2.0.1...v2.0.2
 [v2.0.1]: https://github.com/TechnologyAdvice/Squiss/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/TechnologyAdvice/Squiss/compare/v1.1.0...v2.0.0